refactor:  필요 없는 코드 삭제 및 time필터링 수정
